+++ src/modules/artists/artists.controller.ts
@@ -10,9 +10,9 @@ import {
   Post,
   Put,
 } from '@nestjs/common';
+import Artist from './models/artist.model';
 import CreateArtist from './dto/create-artist.dto';
 import UpdateArtist from './dto/update-artist.dto';
-import Artist from './model/artist.model';
 import { ArtistsService } from './artists.service';
 
 @Controller('artist')
@@ -30,21 +30,21 @@ export class ArtistsController {
   }
 
   @Post()
-  create(@Body() CreateArtistDto: CreateArtist): Promise<Artist> {
-    return this.artistsService.create(CreateArtistDto);
+  create(@Body() input: CreateArtist): Promise<Artist> {
+    return this.artistsService.create(input);
+  }
+
+  @Put(':id')
+  update(
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() input: UpdateArtist,
+  ): Promise<Artist> {
+    return this.artistsService.update(id, input);
   }
 
   @Delete(':id')
   @HttpCode(HttpStatus.NO_CONTENT)
   remove(@Param('id', new ParseUUIDPipe()) id: string): Promise<Artist> {
     return this.artistsService.remove(id);
-  }
-
-  @Put(':id')
-  update(
-    @Body() UpdateArtistDto: UpdateArtist,
-    @Param('id', new ParseUUIDPipe()) id: string,
-  ): Promise<Artist> {
-    return this.artistsService.update(id, UpdateArtistDto);
   }
 }
