+++ src/modules/artists/artists.module.ts
@@ -1,8 +1,10 @@
 import { Module, forwardRef } from '@nestjs/common';
-import { TracksModule } from './../tracks/tracks.module';
+
 import { ArtistsController } from './artists.controller';
 import { ArtistsService } from './artists.service';
+
 import { AlbumsModule } from '../albums/albums.module';
+import { TracksModule } from '../tracks/tracks.module';
 import { FavoritesModule } from '../favorites/favorites.module';
 @Module({
   providers: [ArtistsService],
