+++ src/modules/artists/models/artist.model.ts
