+++ src/modules/favorites/favorites.service.ts
@@ -5,13 +5,15 @@ import {
   UnprocessableEntityException,
 } from '@nestjs/common';
 
+import Album from '../albums/models/album.model';
+import Artist from '../artists/models/artist.model';
+import Track from '../tracks/model/track.model';
+
+import { FavoritesIds, FavoritesResponse } from './model/favorite.model';
+
 import { TracksService } from '../tracks/tracks.service';
 import { AlbumsService } from '../albums/albums.service';
 import { ArtistsService } from '../artists/artists.service';
-import Album from '../albums/model/album.model';
-import Artist from '../artists/model/artist.model';
-import Track from '../tracks/model/track.model';
-import { FavoritesIds, FavoritesResponse } from './model/favorite.model';
 
 @Injectable()
 export class FavoritesService {
